added new classes Keyboard and MixinLang

+++ src/item.py
@@ -26,7 +26,7 @@ class Item:
     def __str__(self):
         return self.__name
 
-    def calculate_total_price(self) -> int:
+    def calculate_total_price(self) -> float:
         """
         Рассчитывает общую стоимость конкретного товара в магазине.
